Javier

+++ backend/obra-social-api/src/main/java/service/AffiliateService.java
@@ -3,14 +3,13 @@ package service;
 import jakarta.enterprise.context.ApplicationScoped;
 import jakarta.inject.Inject;
 import jakarta.transaction.Transactional;
+import jakarta.validation.Valid;
 import model.Affiliate;
 import repository.AffiliateRepository;
 import service.interfaces.IAffiliateService;
-import validator.AffiliateValidator;
+import validator.Validator;
 
 import java.util.List;
-
-import dto.mappers.AffiliateMapper;
 
 @ApplicationScoped
 @Transactional
@@ -20,7 +19,10 @@ public class AffiliateService implements IAffiliateService {
 	private AffiliateRepository affiliateRepository;
 
 	@Inject
-	private AffiliateValidator affiliateValidator;
+	private Validator validator;
+
+	@Inject
+	private ShiftService shiftService;
 
 	@Override
 	public List<Affiliate> getAllAffiliates() {
@@ -36,31 +38,30 @@ public class AffiliateService implements IAffiliateService {
 
 	public Affiliate addAffiliate(Affiliate newAffiliate) throws Exception {
 		if (affiliateRepository.findByDni(newAffiliate.getDni()) != null)
-			throw new Exception("Ya existe un afiliado con el dni: " + newAffiliate.getDni());
+			throw new IllegalArgumentException("Ya existe un afiliado con el dni: " + newAffiliate.getDni());
 		if (affiliateRepository.findByHealthInsuranceCode(newAffiliate.getHealthInsuranceCode()) != null)
-			throw new Exception(
+			throw new IllegalArgumentException(
 					"Ya existe un afiliado con el codigo de obra social: " + newAffiliate.getHealthInsuranceCode());
-		List<String> existingErrors = affiliateValidator.validateAffiliate(AffiliateMapper.entityToDto(newAffiliate));
-		if (existingErrors != null)
-			throw new IllegalArgumentException(existingErrors.toString());
+
+		List<String> existInvalidData = validator.validateAffiliate(newAffiliate);
+		if(existInvalidData!=null) throw new IllegalArgumentException(existInvalidData.toString());
 		affiliateRepository.persist(newAffiliate);
 		return newAffiliate;
 	}
 
 	@Override
-	public Affiliate editAffiliate(Long id, Affiliate editedAffiliate) throws Exception {
+	public Affiliate editAffiliate(Long id, @Valid Affiliate affiliate) throws Exception {
 		Affiliate existingAffiliate = affiliateRepository.findById(id);
 		if (existingAffiliate == null)
 			throw new Exception("El afiliado con id " + id + " no existe.");
-		List<String> existingErrors = affiliateValidator
-				.validateAffiliate(AffiliateMapper.entityToDto(editedAffiliate));
-		if (existingErrors != null)
-			throw new IllegalArgumentException(existingErrors.toString());
+		List<String> existInvalidData = validator.validateAffiliate(affiliate);
+		if(existInvalidData!=null) throw new IllegalArgumentException(existInvalidData.toString());
 
-		existingAffiliate.setFirstName(editedAffiliate.getFirstName());
-		existingAffiliate.setLastName(editedAffiliate.getLastName());
-		existingAffiliate.setDni(editedAffiliate.getDni());
-		existingAffiliate.setHealthInsuranceCode(editedAffiliate.getHealthInsuranceCode());
+		existingAffiliate.setFirstName(affiliate.getFirstName());
+		existingAffiliate.setLastName(affiliate.getLastName());
+		existingAffiliate.setDni(affiliate.getDni());
+		existingAffiliate.setHealthInsuranceCode(affiliate.getHealthInsuranceCode());
+		existingAffiliate.setEmail((affiliate.getEmail()));
 		affiliateRepository.persistAndFlush(existingAffiliate);
 		return existingAffiliate;
 	}
@@ -71,6 +72,7 @@ public class AffiliateService implements IAffiliateService {
 		if (existingAffiliate == null)
 			throw new Exception("El afiliado con id " + id + " no existe.");
 
+		if(!shiftService.getShiftByAffiliateId(id).isEmpty())throw new Exception("No se puede borrar el afiliado con ID: "+id+" debido a que está asociado a un turno.");
 		affiliateRepository.deleteById(id);
 		return existingAffiliate;
 	}
